requirements.txt
numpy
pillow
click
pytest
